% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_barplot}
\alias{add_barplot}
\title{Add a barplot to an interactive scene}
\usage{
add_barplot(scene, encoding = NULL)
}
\arguments{
\item{scene}{A scene object}

\item{encoding}{Encoding of the variables: \code{v1} a discrete variable,
\code{v2} a continuous variable (optional)}
}
\description{
Function that adds a barplot to an interactive scene
}
