% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_bibarplot}
\alias{add_bibarplot}
\title{Add a mirrored barplot to a scene or schema}
\usage{
add_bibarplot(x, variables = NULL, options = NULL)
}
\arguments{
\item{x}{A \code{plotscaper} scene or schema object}

\item{variables}{A vector of variable names:
one discrete (required), one or two continuous (required)}

\item{options}{A list of options}
}
\value{
The scene or schema back, with the plot added appropriately
}
\description{
This function adds a mirrored barplot to a \code{plotscaper} scene or schema.
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
