% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_fluctplot}
\alias{add_fluctplot}
\title{Add a fluctuation diagram to an interactive scene}
\usage{
add_fluctplot(scene, encoding = NULL)
}
\arguments{
\item{scene}{A scene object}

\item{encoding}{Encoding of the variables: \code{v1} and \code{v2} discrete variables,}
}
\description{
Function that adds a fluctuation diagram to an interactive scene
}
