% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_histogram}
\alias{add_histogram}
\title{Add a histogram to an interactive scene}
\usage{
add_histogram(scene, encoding = NULL)
}
\arguments{
\item{scene}{A scene object}

\item{encoding}{Encoding of the variables: \code{v1} a continuous variable}
}
\description{
Function that adds a histogram to an interactive scene
}
