% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_histogram2d}
\alias{add_histogram2d}
\title{Add a 2D histogram to an interactive scene}
\usage{
add_histogram2d(scene, encoding = NULL, options = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object}

\item{encoding}{A vector of variable names:
two continuous (required) and another continuous (optional)}
}
\value{
The scene back
}
\description{
This function adds a 2D histogram to an interactive scene
The squares' areas can be normalized by pressing the normalize ("N") key.
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
