% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_lineplot}
\alias{add_lineplot}
\title{Add a lineplot to an interactive scene}
\usage{
add_lineplot(scene, encoding = NULL, options = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object.}

\item{encoding}{A vector of variable names:
any number of continuous or discrete}
}
\value{
The scene back
}
\description{
This function adds a parallel coordinates plot to an interactive scene
The variables can be put on common scale by pressing the normalize ("N")
key (only works if all continuous).
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
