% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_notes}
\alias{add_notes}
\title{Add a not(e)plot to an interactive scene}
\usage{
add_notes(scene)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object}
}
\value{
The scene back
}
\description{
Function that adds a "noteplot" to an interactive scene. This is not
an actual plot but just a simple HTML \code{textarea} element that you can
use to take notes.
}
