% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_parcoords}
\alias{add_parcoords}
\title{Add a parallel coordinates plot to an interactive scene}
\usage{
add_parcoords(scene, encoding = NULL)
}
\arguments{
\item{scene}{A scene object}

\item{encoding}{Encoding of the variables: \code{v1}, \code{v2}, \code{v3}, ... (continuous or discrete variables)}
}
\description{
Function that adds a parallel coordinates plot to an interactive scene
}
