% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_pcoords}
\alias{add_pcoords}
\title{Add a parallel coordinates plot to a scene or schema}
\usage{
add_pcoords(x, variables = NULL, options = NULL)
}
\arguments{
\item{x}{A \code{plotscaper} scene or schema object}

\item{variables}{A vector of variable names:
at least two continuous or discrete variables}

\item{options}{A list of options}
}
\value{
The scene or schema back, with the plot added appropriately
}
\description{
This function adds a parallel coordinates plot
to a \code{plotscaper} scene or schema.
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
