% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{add_plot.plotscaper_schema}
\alias{add_plot.plotscaper_schema}
\title{Add a plot to a plotscaper scene}
\usage{
\method{add_plot}{plotscaper_schema}(schema, options = NULL)
}
\description{
Add a plot to a plotscaper scene
}
