% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_scatterplot}
\alias{add_scatterplot}
\title{Add a scatterplot to an interactive scene}
\usage{
add_scatterplot(scene, encoding = NULL, options = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object}

\item{encoding}{A vector of variable names:
two continuous or discrete (required), one continuous (optional)}
}
\value{
The scene back
}
\description{
This function adds a scatterplot to an interactive scene.
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
