% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{assign_cases}
\alias{assign_cases}
\title{Assigns cases to a specific group}
\usage{
assign_cases(scene, cases = NULL, group = 1)
}
\arguments{
\item{scene}{A \code{plotscaper} scene}

\item{cases}{The cases (rows) to select}

\item{group}{The group to assign the cases to (can be: \code{1}, \code{2}, or \code{3})}
}
\value{
Nothing (a pure side-effect)
}
\description{
Assigns specific cases (rows of the data) to a permanent group.
Permanent group assignments are only removed by double-clicking.
}
