% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{create_schema}
\alias{create_schema}
\title{Create a \code{plotscaper} schema}
\usage{
create_schema(data = NULL, options = NULL)
}
\arguments{
\item{data}{A dataframe}

\item{options}{A list of options}
}
\value{
An object of class \code{plotscaper_schema}
}
\description{
This function constructs a schema of an interactive
\code{plotscaper} figure.
}
\examples{
create_schema(mtcars) |> add_scatterplot(c("wt", "mpg")) |> render()
}
