% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{dispatch_message}
\alias{dispatch_message}
\title{Dispatches a message to a plotscaper scene or schema}
\usage{
dispatch_message(x, message)
}
\arguments{
\item{x}{A plotscaper scene or schema}

\item{message}{A list that will get converted to JSON message at appropriate time}
}
\value{
The scene or schema back
}
\description{
Dispatches a message to a plotscaper scene or schema
}
