% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{get_plot_ids}
\alias{get_plot_ids}
\title{Return a list of plot ids from a plotscaper scene or schema}
\usage{
get_plot_ids(x)
}
\arguments{
\item{x}{A plotscaper scene or schema}
}
\value{
The scene or schema back
}
\description{
Return a list of plot ids from a plotscaper scene or schema
}
