% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\docType{data}
\name{id}
\alias{id}
\title{Plot id}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
id
}
\description{
A string which uniquely identifies a plot \code{plotscaper} scene or schema.
}
\details{
\code{id} is a string that uniquely identifies a plot  within a
\code{plotscaper} scene or schema. It can match a plot
based on its position (e.g. "plot1", "plot2", ...),
in the order the plots were added, left-to-right top-to-bottom,
or it can match plot based on type (e.g. "scatter1" or "barplot3"),
again, in order of addition.

If the plot is matched based on type, the morphemes "plot" and
"gram" are ignored, such that e.g. "scatterplot1" is the same as
"scatter1" and "histogram2d4" is the same as "histo2d4".

The string can also be shortened, e.g. "p1" for "plot1",
"s2" for "scatter2",  or "hh3" for "histo2d3".
}
\keyword{datasets}
