% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{mark_cases}
\alias{mark_cases}
\title{Mark rows of the data}
\usage{
mark_cases(row_ids = NULL, group = 2)
}
\arguments{
\item{row_ids}{The ids (row numbers) of the cases}

\item{group}{Which group to assign the cases to}
}
\value{
Nothing (a pure side-effect)
}
\description{
Assign specific rows of the data to a given group
within the current plotscaper figure.
}
