% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{normalize}
\alias{normalize}
\title{Normalize a plot}
\usage{
normalize(x, id = NULL)
}
\arguments{
\item{x}{A \code{plotscaper} scene}

\item{id}{A string id of the plot. See \link{id}}
}
\value{
The scene or schema back
}
\description{
This function switches the representation of a plot to a normalized one,
e.g. spineplot, spinogram, etc...
}
