% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{plotscaper_global}
\alias{plotscaper_global}
\title{A Plotscaper Global Configuration Object}
\format{
An object of class \code{environment} of length 6.
}
\usage{
plotscaper_global
}
\description{
Used mainly for setting up the HTTP server for communication
between an interactive R session and the figure.
}
\keyword{datasets}
