% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{pop_plot}
\alias{pop_plot}
\title{Remove the last plot from a scene or schema}
\usage{
pop_plot(x)
}
\arguments{
\item{x}{A plotscaper scene or schema}
}
\value{
The scene or schema back
}
\description{
This function removes the last plot from a \code{plotscaper}
scene or schema.
}
