% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ps_map}
\alias{ps_map}
\title{Create a plotscaper mapping}
\usage{
ps_map(x = NULL, y = NULL, size = NULL)
}
\arguments{
\item{x}{Variable mapped to the x-aesthetic (string)}

\item{y}{Variable mapped to the y-aesthetic (string)}

\item{size}{Variable mapped to the size aesthetic (string)}
}
\description{
A helper function to create an aesthetic mapping for plotscaper plots
}
