% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ps_premark}
\alias{ps_premark}
\title{Pre-mark cases in the data}
\usage{
ps_premark(scene, indices, membership)
}
\arguments{
\item{scene}{A plotscaper scene object}

\item{indices}{A vector of indices indicating which cases in the data should be marked}

\item{mapping}{A number indicating which membership should be assigned to the cases}
}
\description{
Pre-mark cases in the data
}
