% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ps_scatter}
\alias{ps_scatter}
\title{Add a scatterplot to a plotscape scene}
\usage{
ps_scatter(scene, mapping)
}
\arguments{
\item{scene}{A plotscape scene object}

\item{mapping}{A list which declares how variables in the data (belonging to scene) map to visual aesthetics/encodings}
}
\description{
Function that appends a ready-made scatterplot to a plotscape scene. A wrapper for ps_wrapper_plot(scene, type = "scatter", mapping)
}
