% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{ps_scene}
\alias{ps_scene}
\title{Create an Interactive Visual Scene}
\usage{
ps_scene(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data object that can be converted to JSON, preferably in long format}

\item{width}{Width of the visual scene}

\item{height}{Height of the visual scene}
}
\description{
Function that creates a skeleton of an interactive visual scene.
}
\examples{
scene(mtcars, 900, 500)
|> ps_wrapper_plot("scatter", list(x = "wt", y = "mpg"))
|> ps_wrapper_plot("bar", list(x = "cyl", y = "_indicator"))
}
