% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ps_squareheat}
\alias{ps_squareheat}
\title{Add a square heatmap to a plotscaper scene}
\usage{
ps_squareheat(scene, mapping, ...)
}
\arguments{
\item{scene}{A plotscaper scene object}

\item{mapping}{A mapping created with \code{ps_map()} or a list that declares how variables in the data (belonging to scene) map to visual aesthetics/encodings}
}
\description{
Function that appends a ready-made "square heatmap" (with square size representing count, not colour) to a plotscaper scene. A wrapper for ps_wrapper_plot(scene, type = "histo", mapping)
}
