% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ps_wrapper_plot}
\alias{ps_wrapper_plot}
\title{Add a ready-made plot to a plotscape scene}
\usage{
ps_wrapper_plot(scene, type, mapping, height = NULL, width = NULL)
}
\arguments{
\item{scene}{A plotscape scene object}

\item{type}{A string declaring the type of plot to be added, e.g. "scatter", "histo", "bar", "square"}

\item{mapping}{A list which declares how variables in the data (belonging to scene) map to visual aesthetics/encodings}
}
\description{
Function that appends a ready-made plot of a specific type to an existing plotscape scene
}
\examples{
s1 <- scene(mtcars, 900, 500)
s1 |> ps_wrapper_plot(type = "scater", mapping = list(x = "wt", y = "mpg"))
}
