% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{remove_plot}
\alias{remove_plot}
\title{Remove specific plot from a scene or schema}
\usage{
remove_plot(x, id = NULL)
}
\arguments{
\item{x}{A plotscaper scene or schema}

\item{id}{A string id of the plot. See \link{id}}
}
\description{
This function removes a specific plot from a \code{plotscaper}
scene or schema.
}
\examples{
\dontrun{
# you can only run this in interactive R session
scene <- create_schema(mtcars) |>
  add_scatterplot(c("wt", "mpg")) |>
  add_scatterplot(c("disp", "mpg")) |>
  add_barplot(c("cyl")) |>
  render()

scene
scene |> remove_plot("scatter2") # Removes the second scatterplot
scene |> remove_plot("plot2") # Removes the second plot (barplot)
}

}
