% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{render}
\alias{render}
\title{Render a \code{plotscaper} schema}
\usage{
render(schema, width = NULL, height = NULL, elementId = NULL, options = NULL)
}
\arguments{
\item{schema}{A \code{plotscaper} schema object}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Id of the HTML element}
}
\description{
This function takes a \code{plotscaper} schema and renders it as a
concrete \code{htmlwidgets} widget.
}
