% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{reset}
\alias{reset}
\title{Reset a scene or schema}
\usage{
reset(x)
}
\arguments{
\item{x}{A \code{plotscaper} scene or schema}
}
\value{
The scene or schema back
}
\description{
This function resets a \code{plotscaper} scene or schema.
All selection/group assignment will be removed, and
axis limits/levels of zoom will be restored to default.
}
