% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{select_cases}
\alias{select_cases}
\title{Select cases}
\usage{
select_cases(scene, cases = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene}

\item{cases}{The cases (rows) to select}
}
\value{
Nothing (a pure side-effect)
}
\description{
Select specific cases (rows of the data) by assigning them
to transient selection. Note that transient selection is removed when
any plot is clicked. To make the selection more permanent, use \code{assign_cases()}.
}
