% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{selected_cases}
\alias{selected_cases}
\title{Check selected rows of the data}
\usage{
selected_cases(group = 1:8, resolvefn = print)
}
\arguments{
\item{group}{Which group to assign the cases to}

\item{resolvefn}{Function to call after the server receives a response from
the figure. Defaults to \code{print}}
}
\value{
Nothing (use \code{resolvefn} to save output)
}
\description{
Check which cases of the data are selected
within the current plotscaper figure.
}
\examples{
# Assign selected cases from group 2 to the `cases` variable
# selected_cases(2, resolvefn = function(x) cases <<- x)
}
