% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{set_layout}
\alias{set_layout}
\title{Set interactive scene layout}
\usage{
set_layout(scene = NULL, layout = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene}

\item{layout}{A numeric matrix of contiguous rectangle ids}
}
\value{
The scene back
}
\description{
This function sets a layout for a \code{plotscaper} scene, similar to the
\code{graphics::layout} function.
}
