% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{set_parameters}
\alias{set_parameters}
\title{Set reactive parameters}
\usage{
set_parameters(
  x,
  id = NULL,
  width = NULL,
  anchor = NULL,
  width_x = NULL,
  anchor_x = NULL,
  width_y = NULL,
  anchor_y = NULL
)
}
\arguments{
\item{x}{A \code{plotscaper} scene}

\item{id}{A string id of the plot. See \link{id}}

\item{width}{Histogram binwidth}

\item{anchor}{Histogram anchor}

\item{width_x}{2D histogram binwidth (x-axis)}

\item{anchor_x}{2D histogram anchor (x-axis)}

\item{width_y}{2D histogram binwidth (y-axis)}

\item{anchor_y}{2D histogram anchor (y-axis)}
}
\value{
The scene or schema back
}
\description{
This functions sets reactive paramaters on a plot such as a histogram.
}
