% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{set_scale}
\alias{set_scale}
\title{Set values of a scale}
\usage{
set_scale(
  x,
  id = NULL,
  scale = NULL,
  min = NULL,
  max = NULL,
  breaks = NULL,
  direction = NULL,
  mult = NULL,
  default = NULL
)
}
\arguments{
\item{x}{A \code{plotscaper} scene or schema}

\item{id}{A string id of the plot. See \link{id}}

\item{scale}{A string identifying scale. Can be: "x", "y", "area", or "size".}

\item{min}{Scale minimum. Only works for continuous scales}

\item{max}{Scale maximum. Only works for continuous scales}

\item{breaks}{A vector of scale breaks. Only works for discrete scales.
The values should be the same as in the original scale, just different order.}

\item{direction}{Scale direction. Can be \code{1} or \code{-1}}

\item{mult}{Scale multiplier}
}
\description{
This function sets the values of a scale within one plot
inside a \code{plotscaper} scene or schema.
}
