% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{set_scene}
\alias{set_scene}
\title{Set up an interactive scene}
\usage{
set_scene(
  data = NULL,
  options = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A dataframe that will be converted to JSON
(missing values are currently not supported).}

\item{options}{A list of options}

\item{width}{Width of the scene}

\item{height}{Height of the scene}

\item{elementId}{An id of the element to render the widget in}
}
\value{
An object of class \code{htmlwidget}
}
\description{
This function constructs a skeleton of an interactive
\code{plotscaper} figure. Specifically, it parses the data to JSON,
creates an object of class \code{htmlwidget} and sends it
additional information that is necessary to set up the figure,
such as the variable types,
the URL of the websocket server (if interactive), etc..
The figure gets rendered when the resulting \code{htmlwidget}
object gets printed.
}
\examples{
set_scene(mtcars) |> add_scatterplot(c("wt", "mpg"))
}
