% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{start_server}
\alias{start_server}
\title{Start a server for interactive figure use}
\usage{
start_server(random_port = FALSE)
}
\arguments{
\item{random_port}{Whether to use a random port number. Useful if default port is already taken.}
}
\value{
Nothing (a pure side-effect)
}
\description{
Starts an httpuv server for an interactive communication with the plotscaper figure.
Relies on \code{plotscaper_global} options.
}
