% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_fluctplot}
\alias{add_fluctplot}
\title{Add a fluctuation diagram to an interactive scene}
\usage{
add_fluctplot(scene, encoding = NULL, options = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object}

\item{encoding}{A vector of variable names:
two discrete (required) and one continuous (optional)}

\item{options}{A list of options}
}
\value{
The scene back
}
\description{
This function adds a fluctuation diagram to an interactive scene.
The squares' areas can be normalized by pressing the normalize ("N") key.
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
