% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_histogram}
\alias{add_histogram}
\title{Add a histogram to an interactive scene}
\usage{
add_histogram(scene, encoding = NULL, options = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object.}

\item{encoding}{Encoding of the variables:
\code{v1} a continuous variable, \code{v2} a continuous variable (optional).}

\item{options}{A list of options.}
}
\description{
This function adds a histogram to an interactive scene.
Can be transformed into spinogram by pressing the normalize ("N") key.
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
