% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_histogram2d}
\alias{add_histogram2d}
\title{Add a 2D histogram to an interactive scene}
\usage{
add_histogram2d(scene, encoding = NULL)
}
\arguments{
\item{scene}{A scene object}

\item{encoding}{Encoding of the variables: \code{v1} and \code{v2} continuous}
}
\description{
Function that adds a 2D histogram to an interactive scene
}
