% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_plot}
\alias{add_plot}
\title{Add plot to an interactive scene}
\usage{
add_plot(scene, type = NULL, encoding = NULL, options = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object}

\item{type}{A string representing the plot type}

\item{encoding}{A vector of variable encodings}

\item{options}{A list of options}
}
\value{
The scene back (with the added specification for the plot)
}
\description{
This function adds the specification for a additional plot to
an existing \code{plotscaper} scene (\code{htmlwidget}) object. The type of plot is
specified via a string. Variable encodings are specified via a vector
of character strings, corresponding to the names of the variables in
the data.
}
\details{
This function is used internally in specialized wrapper functions such as \link{add_scatterplot}
or \link{add_barplot}, and I recommend to use these functions instead.

Note that the variable encodings specified via the \code{encoding} argument
are different from aesthetic encodings used in Grammar of Graphics
(GoG)-based systems such as \code{ggplot2}.
That is, code like the following:

\if{html}{\out{<div class="sourceCode">}}\preformatted{add_plot("scatter", c(x = "wt", y = "mpg"))
}\if{html}{\out{</div>}}

will not work. This is because, in \code{plotscaper},
data variables are mapped to intermediate "summary" variables which may
be subject to some statistical transformations,
and only after that happens can these summary variables
be mapped to aesthetic variables. I.e. there is a three-step model:

\deqn{\textbf{data}
\overset{\text{summarize}}{\longrightarrow} \textbf{summaries}
\overset{\text{encode}}{\longrightarrow} \textbf{aesthetics}}

The reason for this departure from the GoG style of specifying aesthetics
is that some interactive plots are difficult to express
as a clean 1-to-1 mapping from data to aesthetics.

For example, in a spinogram (a scaled version
of histogram - click on a histogram and press the "N" key),
the y-axis variable is
scaled count, and x-axis variable is cumulative (absolute) count.
Neither of these variables come directly from the data, and both
represent a transformation of the same underlying summary variable
(count within bins). In fact, the underlying summary variable in a
spinogram is exactly the same as the one histogram uses - it is
just transformed differently and mapped to different aesthetics in
each kind of plot.
}
