% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{add_plot}
\alias{add_plot}
\title{Add a plot to a scene or schema}
\usage{
add_plot(x, spec)
}
\arguments{
\item{x}{A plotscaper scene or schema}

\item{spec}{A list with the plot specification}
}
\value{
The scene or schema back
}
\description{
This function adds a plot to an existing \code{plotscaper}
scene or schema. Not meant to be called directly
but instead with a wrapper function such as
\code{\link[=add_scatterplot]{add_scatterplot()}}.
}
