% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_scatterplot}
\alias{add_scatterplot}
\title{Add a scatterplot to a scene or schema}
\usage{
add_scatterplot(x, variables = NULL, options = NULL)
}
\arguments{
\item{x}{A \code{plotscaper} scene or schema object}

\item{variables}{A vector of variable names:
two continuous or discrete (required), one continuous (optional)}

\item{options}{A list of options}
}
\value{
The scene or schema back, with the plot added appropriately
}
\description{
This function adds a scatterplot to a \code{plotscaper} scene or schema.
}
\seealso{
\code{\link[=add_plot]{add_plot()}}
}
