% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{assign_cases}
\alias{assign_cases}
\title{Assign cases to a group}
\usage{
assign_cases(x, cases = NULL, group = 1)
}
\arguments{
\item{x}{A \code{plotscaper} scene or schema}

\item{cases}{The cases (rows) to select}

\item{group}{The group to assign the cases to (can be 1, 2, or 3)}
}
\value{
The scene or schema back
}
\description{
This function assigns specific cases (rows of the data)
to a permanent group within a \code{plotscaper} scene or schema.
Permanent group assignments are only removed by double-clicking.
}
