% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{clear_layout}
\alias{clear_layout}
\title{Set interactive scene layout}
\usage{
clear_layout(x)
}
\arguments{
\item{x}{A \code{plotscaper} scene}
}
\value{
The scene or schema back
}
\description{
This function clears an existing layout. See \code{\link[=set_layout]{set_layout()}}.
}
