% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducer.R
\name{reducer}
\alias{reducer}
\title{Create a reducer}
\usage{
reducer(initialfn = NULL, reducefn = NULL, name = NULL)
}
\arguments{
\item{reducefn}{A reducer function.}

\item{name}{A name for the reducer}

\item{init}{An initializing function.}
}
\description{
Create a reducer
}
