% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{remove_plot}
\alias{remove_plot}
\title{Remove specific plot from a scene or schema}
\usage{
remove_plot(x, id = NULL)
}
\arguments{
\item{x}{A plotscaper scene or schema}

\item{id}{A string id of the plot. See \link{id}}
}
\value{
The scene or schema back
}
\description{
This function removes a specific plot from a \code{plotscaper}
scene or schema.
}
