% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{render}
\alias{render}
\title{Render a schema into an interactive scene}
\usage{
render(
  schema,
  launch_server = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL,
  options = NULL
)
}
\arguments{
\item{schema}{A \code{plotscaper} schema object}

\item{launch_server}{Whether to launch a httpuv server for interaction with figure}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Id of the HTML element to render the scene in (optional)}

\item{options}{A list of options}
}
\value{
An object of class \code{plotscaper_scene}
}
\description{
This function takes a \code{plotscaper} schema and renders it as a
concrete \code{htmlwidgets} widget.
}
