% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{select_cases}
\alias{select_cases}
\title{Select cases of the data}
\usage{
select_cases(x, cases = NULL)
}
\arguments{
\item{x}{A \code{plotscaper} scene or schema}

\item{cases}{The cases (rows) to select}
}
\value{
The scene or schema back
}
\description{
This function selects specific cases (rows of the data)
within a \code{plotscaper} scene or schema by assigning
them to transient selection.
Transient group assignment is removed by clicking.
}
