% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{set_layout}
\alias{set_layout}
\title{Set interactive scene layout}
\usage{
set_layout(scene = NULL, layout = NULL)
}
\arguments{
\item{scene}{A \code{plotscaper} scene object.}
}
\description{
This function sets a layout for the interactive scene, similar to the
\code{graphics::layout} function.
}
