% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_senders.R
\name{set_layout}
\alias{set_layout}
\title{Set interactive scene layout}
\usage{
set_layout(x, layout = NULL)
}
\arguments{
\item{x}{A \code{plotscaper} scene}

\item{layout}{A numeric matrix of plot ids, arranged into contiguous rectangles}
}
\value{
The scene or schema back
}
\description{
This function sets a layout for a \code{plotscaper} scene. Similar to the
\code{graphics::layout} function.
}
