% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{set_scene}
\alias{set_scene}
\title{Set up an interactive scene}
\usage{
set_scene(data = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A dataframe that will be converted to JSON
(missing values are not currently supported).}

\item{width}{Width of the scene}

\item{height}{Height of the scene}
}
\description{
This function constructs the underlying skeleton of an interactive
\code{plotscaper} scene. Specifically, it parses the data, sends it to
to \code{plotscape} via \code{htmlwidgets}, and sets up the scene object which
takes care of adding plots and between-plot interactions.
}
\examples{
set_scene(mtcars) |> add_scatterplot(c("wt", "mpg"))
}
